Add files via upload
Created the file that creates and instantiates the Database class used to access our database. It has a constructor and a simple query execution function. The query execution function takes a query template, the fields that have to be prepared into the template (optional, as it can be used to executed a simple query) and the return mode (optional, by default it's set to an associative array).
Created the file that contains the info used by the database object to connect to the MySQL database. For now, they are set to temporary local test values.